Story/1.emotions interactive list (#3)
* Add basic meta tags to head of pages

* Create basic dynamic list

Adds a list that the user can interact with by clicking. The list
currently does not have dynamic text.

* Change emotion insert order

Inserts a new emotion after the placeholder instead of at the end
of the list.

* Add gitignore

* Switch to Flask development

* Add link on register to login page

* Add basic CSS

* Add basic db structure

* Restructure flask app

* Add basic registration call to db

* Add database files to ignore

* Add login verification and routing

* Add emotions page and cbt home page

* Update CBT log pages to use Jinja

* Split repeated code into a jinja2 macro

* Start restructuring log pages under one endpoint

* Change navbar for 'account' page

Make the link for 'account' change to 'home' if you're
on the 'account' page.

* Add datetime default to event form

* Change routing method for javascript

* Add 'home' to navbar

* Add page redirection for event submit

* Update User model: no image, getid

The User model doesn't need an image at this point, and also it's
useful to have a `getid` function to get the current_user id.

* Create a CBTLog on Submit

Actually enter data into the database, and send the log id num to the
next step in the process.

* Give ability to edit CBTLog

For now, remove datetime from model for simplicity. Give the ability
to go back and view/update a CBTLog in the `/log` endpoint.

* Update account page

* Ensure users can't see each other's logs

* List CBTLogs on home page

* Add edit option to log displays

* Fix bug that doesn't update logs on edit

* Make emotions list creation better

Utilitze the example from w3schools on a todo list for the emotions
list. Include the slider and a delete button for mistakes.

* Fix emotions text not getting pulled

* "Enter" in emotions text input adds item

Enable hitting "enter" to add emotion instead of having to click on the
"Add" button. Also, fix the issue that was causing the blank emotion
warning to not pop up.

* Fix layout and interaction for emotions

Fix the way the loading of the script is done (put in head instead of
body). Also add some more CSS spacing for the emotions list elements,
and remove the javascript from inside the HTML source.